+++ public/locales/zh-CN/game-summonerwars.json
@@ -689,6 +689,54 @@
       "name": "血腥狂怒衰减",
       "description": "在你的回合结束时，从本单位上移除 2 点充能。"
     },
+    "mogu_blood_bloom": {
+      "name": "血腥绽放",
+      "description": "当本召唤师 2 个区格以内的一个友方单位被消灭后，将本召唤师 2 个区格以内的其他友方单位充能。"
+    },
+    "mogu_blood_rage": {
+      "name": "血腥狂怒",
+      "description": "每当一个单位在你的回合中被消灭时，将本单位充能。"
+    },
+    "mogu_blood_rage_decay": {
+      "name": "血腥狂怒衰减",
+      "description": "在你的回合结束时，从本单位上移除至多 2 点充能。"
+    },
+    "mogu_final_form": {
+      "name": "最终形态",
+      "description": "在召唤本单位时，如果你控制一个至少有 5 点充能的菌化野兽，则消灭该单位，并将本单位放置到其所在区格。"
+    },
+    "mogu_blood_infusion": {
+      "name": "鲜血灌注",
+      "description": "每回合一次，在你的移动阶段，指定本单位 2 个区格以内的一个友方单位为目标。将目标充能，然后对其造成 1 点伤害。"
+    },
+    "mogu_transmission": {
+      "name": "菌落传导",
+      "description": "每回合一次，在本单位移动之后，你可以将本单位或其 2 个区格以内一个友方单位上的任意数量充能，移动到本单位 2 个区格以内的另一个友方单位上。"
+    },
+    "mogu_decay": {
+      "name": "腐坏",
+      "description": "在你的移动阶段结束时，对本单位造成 1 点伤害。若本单位存活，将其相邻的一个友方单位充能 2 次。"
+    },
+    "mogu_infection": {
+      "name": "感染",
+      "description": "在本单位消灭一个单位之后，你可以使用弃牌堆中的一个菌袍疫病体替换被消灭的单位。"
+    },
+    "mogu_parasite": {
+      "name": "寄生",
+      "description": "在你的攻击阶段结束时，从本单位上移除 1 点充能；如果本单位没有充能，则改为对其造成 1 点伤害。"
+    },
+    "mogu_burst": {
+      "name": "爆裂",
+      "description": "在你的魔力阶段结束时，如果本单位至少有 3 点充能，则将其消灭。"
+    },
+    "mogu_fungal_mutation": {
+      "name": "菌化变异",
+      "description": "当至少有 3 点充能的本单位被消灭时，你可以使用弃牌堆中的一个菌化野兽替换它。"
+    },
+    "mogu_fanatical_fungus": {
+      "name": "狂热菌菇",
+      "description": "持续。在你移动一个单位之后，可以将其充能。如果你这样做，首先可以将其推拉 1 个区格，然后必须对其造成 1 点伤害。"
+    },
     "sacrifice": {
       "name": "献祭",
       "description": "在本单位被消灭之后，对其相邻的每个敌方单位造成 1 点伤害。"
