+++ src/games/summonerwars/ui/StatusBanners.tsx
@@ -23,8 +23,7 @@ import type {
   WithdrawModeState,
   TelekinesisTargetModeState,
 } from './modeTypes';
-
-type BannerTranslate = (key: string, options?: Record<string, unknown>) => string;
+import { getAbilityModeBannerFallbackText } from './statusBannerText';
 
 // ============================================================================
 // 类型定义
@@ -52,27 +51,6 @@ export interface AnnihilateModeState {
 export interface FuneralPyreModeState {
   cardId: string;
   charges: number;
-}
-
-function getAbilityModeBannerFallbackText(
-  t: BannerTranslate,
-  abilityMode: AbilityModeState,
-): string {
-  if (abilityMode.abilityId === 'fortress_power' && abilityMode.step === 'selectCard') {
-    return t('cardSelector.fortressPower');
-  }
-
-  if (abilityMode.abilityId === 'telekinesis_instead' && abilityMode.step === 'selectUnit') {
-    const abilityName = t('statusBanners.abilityNames.telekinesis_instead');
-    return t('statusBanners.afterAttack.message', { ability: abilityName });
-  }
-
-  if (abilityMode.abilityId === 'high_telekinesis_instead' && abilityMode.step === 'selectUnit') {
-    const abilityName = t('statusBanners.abilityNames.high_telekinesis_instead');
-    return t('statusBanners.afterAttack.message', { ability: abilityName });
-  }
-
-  return '';
 }
 
 // ============================================================================
